OSNF Everywhere: strict outermost-shift normal form
Enforce the invariant that every compound expression has fvar_lb=0,
with Shift(n, core) as the only representation for open expressions.
This ensures pointer equality is reliable for shifted variants of the
same structural expression.

Key changes:
- Remove cutoff from Expr::Shift, make amount u16
- mk_app/mk_pi/mk_lambda/mk_let/mk_proj normalize to OSNF when fvar_lb>0
- adjust_child_tc: O(1) adjustment of OSNF-normalized children
- view_expr transparently sees through Shift nodes
- Systematic read_expr → view_expr migration in inductive.rs, nanoda_tc.rs,
  tc.rs for all pattern matches on Pi/Lambda/App/Proj constructors
- Fix whnf infinite loop: push_shift_up on a Pi can return the same Shift
  due to OSNF re-normalization; use view_expr instead of read_expr+push
- Fix def_eq binder comparison: Shift-wrapped Pi/Lambda pairs need
  view_expr_pair to match correctly
- Parser already enforces OSNF during parsing

All 126 tutorial tests pass.

+++ src/osnf.rs
@@ -189,7 +189,7 @@ fn core_hash(
         | Expr::Local { .. } => expr.get_hash(),
         Expr::Shift { inner, amount, .. } => {
             // OSNF Shift nodes from parse-time normalization: hash the inner core shifted
-            core_hash(dag, inner.idx(), shift + amount as u16, cutoff, memo)
+            core_hash(dag, inner.idx(), shift + amount, cutoff, memo)
         }
     };
 
@@ -412,7 +412,7 @@ fn compute_core(
         | Expr::StringLit { .. } | Expr::Local { .. } => (idx, false),
         Expr::Shift { inner, amount, .. } => {
             // OSNF Shift nodes from parse-time normalization: recurse into inner with combined shift
-            let (core_idx, was_new) = compute_core(dag, inner.idx(), shift + amount as u16, cutoff, memo);
+            let (core_idx, was_new) = compute_core(dag, inner.idx(), shift + amount, cutoff, memo);
             (core_idx, was_new)
         }
     };
